+++ src/main/java/com/slcube/shop/business/address/domain/Address.java
@@ -1,6 +1,6 @@
 package com.slcube.shop.business.address.domain;
 
-import com.slcube.shop.business.member.Member;
+import com.slcube.shop.business.member.domain.Member;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
